pyproject.toml
[project]
name = "ast-grep-mcp"
version = "0.1.0"
description = "MCP server using ast-grep for code analysis and refactoring"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "AST-Grep MCP Contributors", email = "info@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["ast", "grep", "code", "analysis", "refactoring", "mcp"]
license = {text = "MIT"}
dependencies = [
    "ast_grep_py==0.37.0",
    "fastmcp>=0.1.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
ast-grep-mcp = "ast_grep_mcp.__main__:main_entry_point"

[project.urls]
Homepage = "https://github.com/yourorg/ast-grep-mcp"
Documentation = "https://github.com/yourorg/ast-grep-mcp/docs"
Repository = "https://github.com/yourorg/ast-grep-mcp.git"
Issues = "https://github.com/yourorg/ast-grep-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
extra-languages = [
    "tree-sitter-lua>=0.0.14",
    "tree-sitter-go>=0.20.0",
]

[tool.pytest.ini_options]
testpaths = ["tests", "integration_tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "benchmark: marks tests as benchmark tests (deselect with '-m \"not benchmark\"')"
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true
explicit_package_bases = true
mypy_path = "$MYPY_CONFIG_FILE_DIR"
